split page into modular tab components
