+++ frontend/src/app/pages/teacher/hp-system/CohortDetails.tsx
